Fix: Do not overwrite args.target on Windows if value is set
Fixes the unexpected behaviour on Windows that causes the args.target to
always be overwritten. This effectively disables the --target argument on
Windows, thus requiring manual edits to the constants to change the
target destination path.

This commit also changes the default behaviour on non-Windows systems,
so that default values are no longer being strictly enforced when additional
or explicit parameter values have been given for the --target argument.